Move get_joint_MAPQ to unused.cpp
